pyproject.toml
[project]
name = "tarkovdiscbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "hikari>=2.5.0",
    "hikari-arc>=2.1.1",
]
